requirements.txt
yt-dlp==2023.11.16
openai-whisper==20231117
tqdm==4.66.1
